Rename schema files and update imports
